+++ packages/core/test/test-util/request-mocker.ts
@@ -175,7 +175,7 @@ export function mockHeaderRequest({
   path
 }: MockHeaderRequestParams) {
   return nock(host)
-    .get(path ? `${request.serviceUrl()}/${path}` : request.serviceUrl())
+    .head(path ? `${request.serviceUrl()}/${path}` : request.serviceUrl())
     .reply(200, undefined, responseHeaders);
 }