+++ packages/core/v4.d.ts
@@ -1,3 +1,0 @@
-// eslint-disable-next-line import/no-internal-modules
-export * from './dist/v4';
-// # sourceMappingURL=v4.d.ts.map